refactor(input): 입력 서비스 삭제
작업 내용
- 입력 뷰를 제어하는 `InputService`를 삭제하고, 해당 서비스의 로직을 뷰로 이동
