+++ src/main/java/lotto/config/ApplicationConfiguration.java
@@ -1,6 +1,5 @@
 package lotto.config;
 
-import lotto.service.InputService;
 import lotto.service.LottoService;
 import lotto.view.InputView;
 import lotto.view.OutputView;
@@ -12,11 +11,10 @@ public class ApplicationConfiguration {
     private final Scanner scanner = new Scanner(System.in);
     private final InputView inputView = new InputView(scanner);
     private final OutputView outputView = new OutputView();
-    private final InputService inputService = new InputService(inputView, outputView);
     private final LottoService lottoService = new LottoService();
 
-    public InputService getInputService() {
-        return inputService;
+    public InputView getInputView() {
+        return inputView;
     }
 
     public OutputView getOutputView() {
